Make zoom snappier
git-svn-id: svn+ssh://solsvn.oma.be/home/sidc/ssr/SWHV_repos/trunk/client/jhv-3d-wcs@7302 b4e469a2-07ce-4b26-9273-4d7d95a670c7

+++ src/jhv/src/org/helioviewer/jhv/camera/Camera.java
@@ -120,7 +120,7 @@ public class Camera {
         return cameraWidth;
     }
 
-    public void zoom(int wr) {
+    public void zoom(double wr) {
         setCameraFOV(fov * (1 + 0.015 * wr));
     }
